# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave requests
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveListResponse"
    post:
      operationId: createLeave
      tags: [leaves]
      summary: Create leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
  /leaves/{id}:
    get:
      operationId: getLeave
      tags: [leaves]
      summary: Get leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
    put:
      operationId: updateLeave
      tags: [leaves]
      summary: Update leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaveBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
    delete:
      operationId: deleteLeave
      tags: [leaves]
      summary: Delete leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CUSTOMERS ───
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # EMPLOYEE
...
    Leave:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        leaveType:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        days:
          type: number
        reason:
          type: string
        status:
          type: string
        approvedBy:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, employeeId, employeeName, leaveType, startDate, endDate, days, reason, status, createdAt]
    CreateLeaveBody:
      type: object
      properties:
        employeeId:
          type: integer
        leaveType:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
      required: [employeeId, leaveType, startDate, endDate, reason]
    UpdateLeaveBody:
      type: object
      properties:
        status:
          type: string
        approvedBy:
          type: string
    LeaveListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Leave"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]

    # CUSTOMER
...